Add a regression test and organize tests.
