+++ src/test/compile-fail/directory_ownership/macro-expanded-mod.rs
