+++ src/test/compile-fail/directory_ownership/macro_expanded_mod_helper/foo/mod.rs
