+++ src/test/compile-fail/directory_ownership/mod_file_not_owning_aux1.rs
@@ -10,4 +10,7 @@
 
 // ignore-test this is not a test
 
-mod mod_file_not_owning_aux2;
+macro_rules! m {
+    () => { mod mod_file_not_owning_aux2; }
+}
+m!();
