+++ src/test/compile-fail/directory_ownership/mod_file_not_owning_aux2.rs
