+++ src/test/compile-fail/directory_ownership/non-inline-mod-restriction.rs
